api/requirements.txt
affine==2.3.0
attrs==20.3.0
certifi==2020.12.5
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
cmocean==2.0
cycler==0.10.0
Flask==1.1.2
Flask-Cors==3.0.10
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
numpy==1.20.1
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
rasterio==1.2.0
six==1.15.0
snuggs==1.4.7
Werkzeug==1.0.1
